refactor: plugin/toggle
