+++ after/plugin/keymaps.lua
@@ -19,29 +19,6 @@ local fold = require('lib.fold')
 map('n', 'h', fold.h, { desc = 'h: move left or fold' })
 map('n', 'l', fold.l, { desc = 'l: move right and unfold' })
 
--- toggles
--- stylua: ignore start
-local toggle = require('lib.toggle')
-map('n', '<leader>tc', toggle.completion, { desc = 'toggle: completion' })
-map('n', '<leader>ts', toggle.supermaven, { desc = 'toggle: supermaven' })
-map('n', '<leader>tf', toggle.autoformat, { desc = 'toggle: autoformat' })
-Snacks.toggle.option("spell", { name = "Spelling" }):map("<leader>us")
-Snacks.toggle.option("wrap", { name = "Wrap" }):map("<leader>uw")
-Snacks.toggle.option("relativenumber", { name = "Relative Number" }):map("<leader>ur")
-Snacks.toggle.diagnostics():map("<leader>ud")
-Snacks.toggle.line_number():map("<leader>ul")
-Snacks.toggle.option("conceallevel", { off = 0, on = vim.o.conceallevel > 0 and vim.o.conceallevel or 2, name = "Conceal Level" }):map("<leader>uc")
-Snacks.toggle.option("showtabline", { off = 0, on = vim.o.showtabline > 0 and vim.o.showtabline or 2, name = "Tabline" }):map("<leader>uA")
-Snacks.toggle.treesitter():map("<leader>uT")
-Snacks.toggle.option("background", { off = "light", on = "dark" , name = "Dark Background" }):map("<leader>ub")
-Snacks.toggle.dim():map("<leader>uD")
-Snacks.toggle.animate():map("<leader>ua")
-Snacks.toggle.scroll():map("<leader>uS")
-Snacks.toggle.indent():map("<leader>ug")
-Snacks.toggle.profiler():map("<leader>dpp")
-Snacks.toggle.profiler_highlights():map("<leader>dph")
--- stylua: ignore end
-
 map('n', '<leader>R', function()
   vim.cmd(':w')
   vim.cmd(':e! %')
