+++ lazy-lock.json
@@ -1,8 +1,8 @@
 {
-  "SchemaStore.nvim": { "branch": "main", "commit": "806187c790d93625564708c286e16142d87a5cc9" },
+  "SchemaStore.nvim": { "branch": "main", "commit": "8b92ea89835b8e5dbc779a675ebb0e5fcb9a1993" },
   "bamboo.nvim": { "branch": "master", "commit": "1309bc88bffcf1bedc3e84e7fa9004de93da774a" },
   "blink.cmp": { "branch": "main", "commit": "37ce860d8eeeee92d60b9a72db21e369d97e6a75" },
-  "conform.nvim": { "branch": "master", "commit": "9a98b68f8fb289ca8671e9866ba4e0309a693994" },
+  "conform.nvim": { "branch": "master", "commit": "328c2f1dda56f933d3337a1316e1020ba30ba6aa" },
   "dial.nvim": { "branch": "master", "commit": "f97c0c7fa7d5111bc04a91d0f693900fb2d95861" },
   "dropbar.nvim": { "branch": "master", "commit": "ce202248134e3949aac375fd66c28e5207785b10" },
   "everforest-nvim": { "branch": "main", "commit": "557bce922401e247a596583679bc181d4d688554" },
